Streamlit_dashboard/requirements.txt
altair==5.5.0
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
confluent-kafka==2.4.0
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
executing==2.2.0
fonttools==4.56.0
gitdb==4.0.12
gitpython==3.1.44
greenlet==3.1.1
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
jinja2==3.1.5
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-client==8.6.3
jupyter-core==5.7.2
kiwisolver==1.4.8
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
narwhals==1.26.0
nest-asyncio==1.6.0
numpy==2.2.2
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
prompt-toolkit==3.0.50
protobuf==5.29.3
psutil==6.1.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==19.0.0
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.6.1
pydeck==0.9.1
pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32==308
pyzmq==26.2.1
quixstreams==3.8.0
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rocksdict==0.3.25
rpds-py==0.22.3
six==1.17.0
smmap==5.0.2
sqlalchemy==2.0.38
stack-data==0.6.3
streamlit==1.42.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
